Address int warnings in Clownfish core.
Clean up after adding `-Wconversion`.

+++ runtime/core/Clownfish/Class.c
@@ -76,7 +76,7 @@ Class_bootstrap(const cfish_ClassSpec *specs, size_t num_specs,
                                 : offsetof(Class, vtable);
         uint32_t class_alloc_size = novel_offset
                                     + spec->num_novel_meths
-                                      * sizeof(cfish_method_t);
+                                      * (uint32_t)sizeof(cfish_method_t);
 
         Class *klass = (Class*)Memory_wrapped_calloc(class_alloc_size, 1);
 
@@ -134,7 +134,7 @@ Class_bootstrap(const cfish_ClassSpec *specs, size_t num_specs,
                                 : offsetof(Class, vtable);
         uint32_t class_alloc_size = novel_offset
                                     + spec->num_novel_meths
-                                      * sizeof(cfish_method_t);
+                                      * (uint32_t)sizeof(cfish_method_t);
 
         klass->parent           = parent;
         klass->parcel_id        = parcel_id;
@@ -166,7 +166,7 @@ Class_bootstrap(const cfish_ClassSpec *specs, size_t num_specs,
         if (parent) {
             // Copy parent vtable.
             uint32_t parent_vt_size = parent->class_alloc_size
-                                      - offsetof(Class, vtable);
+                                      - (uint32_t)offsetof(Class, vtable);
             memcpy(klass->vtable, parent->vtable, parent_vt_size);
         }
 
@@ -185,7 +185,7 @@ Class_bootstrap(const cfish_ClassSpec *specs, size_t num_specs,
         for (size_t i = 0; i < spec->num_novel_meths; ++i) {
             const NovelMethSpec *mspec = &novel_specs[num_novel++];
             *mspec->offset = novel_offset;
-            novel_offset += sizeof(cfish_method_t);
+            novel_offset += (uint32_t)sizeof(cfish_method_t);
             Class_Override_IMP(klass, mspec->func, *mspec->offset);
         }
     }
